chore: member 엔티티 관련 코드를 user 엔티티 관련 코드로 수정 - 프로필 조회/수정, 이미지 업로드 관련 코드 수정
