--- src/main/java/com/triprint/backend/domain/member/repository/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.triprint.backend.domain.member.repository;
-
-import com.triprint.backend.domain.member.entity.Member;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-	Optional<Member> findByEmail(String email);
-}
