--- src/main/java/com/triprint/backend/domain/member/status/UserRole.java
@@ -1,11 +1,0 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by FernFlower decompiler)
-//
-
-package com.triprint.backend.domain.member.status;
-
-public enum UserRole {
-	ADMIN,
-	USER;
-}
