+++ src/main/java/com/triprint/backend/domain/user/dto/MyProfileResponse.java
@@ -1,4 +1,4 @@
-package com.triprint.backend.domain.member.dto;
+package com.triprint.backend.domain.user.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
