+++ src/main/java/com/triprint/backend/domain/user/dto/kakaoLoginDto/KakaoProfile.java
@@ -1,4 +1,4 @@
-package com.triprint.backend.domain.user.kakaoLoginDto;
+package com.triprint.backend.domain.user.dto.kakaoLoginDto;
 
 import lombok.Data;
